requirements.txt
opencv-python
ultralytics
numpy
torch
torchvision
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.7
langchain-core==1.2.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
opencv-python-headless==4.13.0.90
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
torch
torchvision
ultralytics
streamlit
langchain-openai
python-dotenv
python-box
ensure
matplotlib
seaborn
scikit-learn
scipy
joblib
pyYAML
types-pyYAML
tqdm
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
xxhash==3.6.0
zstandard==0.25.0
joblib
supervision
easyocr
onnxruntime

